document_assistant/pyproject.toml
[tool.poetry]
name = "document-assistant"
version = "0.1.0"
description = "Documentation helper"
authors = ["armandoayala <armandolnayala@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.310"
black = "^23.9.1"
openai = "^0.28.1"
python-dotenv = "^1.0.0"
pinecone-client = "^2.2.4"
tiktoken = "^0.5.1"
pypdf = "^3.16.2"
beautifulsoup4 = "^4.12.2"
unstructured = "^0.10.19"
nltk = "^3.8.1"
fastapi = "^0.103.2"
jinja2 = "^3.1.2"
uvicorn = "^0.23.2"
streamlit = "^1.27.2"
streamlit-chat = "^0.1.1"
tqdm = "^4.66.1"
python-magic-bin = "^0.4.14"

[tool.poetry.scripts]
run-app = "app.main:run_app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
